fix(shader): align diagnostics with Naga validation
- ConstDivideByZero: only flag integer div/mod by zero; float 1.0/0.0 is Inf,
  not an error (matches Naga validate_constant_divisor returning Ok for floats)
- IndexOutOfBounds: bounds-check a constant index against a fixed-size array,
  not just a vector (Naga bounds-checks fixed-size arrays); VariableIdentifier
  now carries arraySize from the symbol's array specifier
- isConstExpr: a macro use site lexes as MACRO_CALL, so its inner node is a
  MacroCallSymbol/Function, not a token — accept it so a #define-sized array
  no longer false-positives NonConstArraySize
- tests: integer vs float div-by-zero AB, array index OOB AB, macro-sized array
  ok, continue-outside/inside-loop MisplacedControlFlow AB

+++ tests/src/shader-analyzer/DiagnosticCoverage.test.ts
@@ -72,7 +72,7 @@ const cases: { code: string; source?: string; gap?: string }[] = [
   },
   {
     code: "ConstDivideByZero",
-    source: pass(`void frag() { float x = 1.0 / 0.0; gl_FragColor = vec4(x); } FragmentShader = frag;`)
+    source: pass(`void frag() { int x = 1 / 0; gl_FragColor = vec4(float(x)); } FragmentShader = frag;`)
   },
   {
     code: "ShiftOutOfRange",
